refactor: move unlocking key class to separate folder
