+++ src/zkscript/types/unlocking_keys/__init__.py
@@ -1,0 +1,5 @@
+"""Classes to generate unlocking keys associated to scripts.
+
+Includes:
+- merkle_tree - implement class MerkleTreeUnlockingKey.
+"""
