+++ tests/merkle_trees/test_merkle_trees.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 import pytest
 from tx_engine import Context
 
-from src.zkscript.merkle_trees.merkle_tree import MerkleTree as MerkleTreeScript
-from src.zkscript.merkle_trees.merkle_tree import MerkleTreeUnlockingKey
+from src.zkscript.merkle_tree.merkle_tree import MerkleTree as MerkleTreeScript
+from src.zkscript.types.unlocking_keys.merkle_tree import MerkleTreeUnlockingKey
 from tests.fields.util import save_scripts
